// components/Unit/0.3.0.4/src/ts/layer3/GitTextIORScenario.interface.ts
/**
 * SPDX-License-Identifier: AGPL-3.0-only WITH AI-GPL-Addendum
 * SPDX-FileComment: See ../../../../../../AI-GPL.md for AI-specific terms.
 * Copyright (c) 2025 Cerulean Circle GmbH TestVeränderung
 * Copyleft: See AGPLv3 (../../../../../../LICENSE) and AI-GPL Addendum (../../../../../../AI-GPL.md)
 * Backlinks: /LICENSE, /AI-GPL.md
 */

import { GitPositioning } from './GitPositioning.interface.js';

export interface GitTextIORScenario {
  ior: {
    uuid: string;
    component: string;
    version: string;
  };
  owner: string;
  model: {
    uuid: string;
    gitUrl: string;
    iorFormat: string;
    positioning: GitPositioning;
    createdAt: string;
    updatedAt: string;
  };
}